Placed example code
