--- .env.example
@@ -1,1 +1,0 @@
-APP_UNTERNET_API_KEY="<your-api-key>"
