--- kernel/example/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@unternet/client-kernel",
-  "version": "0.1.5",
-  "scripts": {
-    "start": "tsx src/index.ts"
-  },
-  "dependencies": {
-    "@ai-sdk/anthropic": "^1.2.10",
-    "@unternet/kernel": "file:..",
-    "chalk": "^5.4.1",
-    "mime-types": "^3.0.1",
-    "readline": "^1.3.0",
-    "shell-quote": "^1.8.2",
-    "untildify": "^5.0.0"
-  },
-  "devDependencies": {
-    "dotenv": "^16.4.7",
-    "tsx": "^4.19.3"
-  }
-}
