--- kernel/example/src/commands/index.ts
@@ -1,1 +1,0 @@
-export * as file from './file';
