--- kernel/example/src/types.ts
@@ -1,1 +1,0 @@
-export type Command = 'exit' | 'file';
