--- kernel/src/actions/processes.ts
@@ -1,4 +1,0 @@
-interface Process {
-  pid: string;
-  uri: string;
-}
