--- kernel/src/interpreter/model.ts
@@ -1,1 +1,0 @@
-export { LanguageModel } from 'ai';
