+++ kernel/src/protocols.ts
@@ -1,21 +1,23 @@
-import { ActionDirective, Protocol } from '../../src';
+import { ActionDirective, Protocol } from '@unternet/kernel';
 
 const protocols: Protocol[] = [
   {
     scheme: 'function',
-    handler: (directive: ActionDirective) => {
+    handleAction: (directive: ActionDirective) => {
       if (directive.actionId === 'get_weather') {
-        return {
+        const content = {
           high: '65F',
           low: '55F',
           summary:
             'Partly cloudy with a chance of light rain in the afternoon.',
         };
+        return { content };
       } else if (directive.actionId === 'check_traffic') {
-        return {
+        const content = {
           conditions: 'Traffic is clear',
           additionalWaitTime: 0,
         };
+        return { content };
       }
     },
   },
