--- kernel/src/shared/types.ts
@@ -1,20 +1,0 @@
-export { Schema } from 'ai';
-
-export interface JSONSchemaDefinition {
-  type:
-    | 'object'
-    | 'string'
-    | 'number'
-    | 'integer'
-    | 'array'
-    | 'boolean'
-    | 'null';
-  description?: string;
-  properties?: {
-    [key: string]: JSONSchemaDefinition;
-  };
-  required?: string[];
-  additionalProperties?: boolean;
-}
-
-export type StringEnum = [string, ...string[]];
