+++ kernel/src/types.ts
@@ -1,0 +1,1 @@
+export type Command = 'exit' | 'file';
