--- kernel/tests/common/tooling.ts
@@ -1,1 +1,0 @@
-export * from 'vitest';
